# static/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "httpcan.org",
...
    "version": "0.7.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://httpcan.org",
      "description": "Demo server"
    }
  ],
...
  "paths": {
...
    "/oauth2/token": {
      "post": {
        "tags": [
          "OAuth 2.0 Mock"
        ],
        "summary": "Token endpoint (four grants)",
        "operationId": "createOauth2Token",
        "description": "RFC 6749 token endpoint supporting authorization_code (+PKCE), password (legacy), refresh_token (rotating), and client_credentials. Client authentication via Authorization: Basic or form-body client_secret; both at once is rejected. Responses carry Cache-Control: no-store and Pragma: no-cache (§5.1). Without --oauth2-clients any non-empty secret passes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "grant_type"
                ],
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": [
                      "authorization_code",
                      "password",
                      "refresh_token",
                      "client_credentials"
                    ]
                  },
                  "code": {
                    "type": "string"
                  },
                  "redirect_uri": {
                    "type": "string"
                  },
                  "code_verifier": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "refresh_token": {
                    "type": "string"
                  },
                  "scope": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token response: access_token, token_type, expires_in, optional refresh_token and scope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request / invalid_grant / unsupported_grant_type (RFC 6749 §5.2 JSON)"
          },
          "401": {
            "description": "invalid_client (when --oauth2-clients is configured)"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "digestAuth": {
        "type": "http",
        "scheme": "digest"
      }
    }
  },
...
}